homologador/requirements.txt
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6-sip==13.6.0
argon2-cffi==23.1.0
portalocker==2.8.2
python-dateutil==2.8.2
pandas==2.2.2
openpyxl==3.1.2